+++ inject/inject-thrift-client/src/test/scala/com/twitter/inject/thrift/integration/serviceperendpoint/GreeterThriftMethodBuilderClientModule.scala
@@ -19,7 +19,7 @@ object GreeterThriftMethodBuilderClientModule
   override val dest = "flag!greeter-thrift-service"
   override val label = "greeter-thrift-client"
 
-  override def configureServicePerEndpoint(
+  override protected def configureServicePerEndpoint(
     builder: ThriftMethodBuilderFactory[Greeter.ServicePerEndpoint],
     servicePerEndpoint: Greeter.ServicePerEndpoint
   ): Greeter.ServicePerEndpoint = {